Renaming sk-python-flask to sk-python-flask-chatgpt-plugin
