+++ sk-python-flask-chatgpt-plugin/.vscode/launch.json
@@ -11,7 +11,7 @@
             "request": "launch",
             "module": "flask",
             "env": {
-                "FLASK_APP": "sk_python_flask/app.py",
+                "FLASK_APP": "sk_python_flask_chatgpt_plugin/app.py",
                 "FLASK_DEBUG": "1"
             },
             "args": [
