+++ sk-python-flask-chatgpt-plugin/skills/FunSkill/Excuses/config.json
