+++ sk-python-flask-chatgpt-plugin/skills/FunSkill/Excuses/skprompt.txt
