+++ sk-python-flask-chatgpt-plugin/skills/FunSkill/Joke/config.json
