+++ sk-python-flask-chatgpt-plugin/skills/FunSkill/Joke/skprompt.txt
