+++ sk-python-flask-chatgpt-plugin/skills/FunSkill/Limerick/config.json
