+++ sk-python-flask-chatgpt-plugin/skills/FunSkill/Limerick/skprompt.txt
